server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.1.1
SQLAlchemy==2.0.23
mysql-connector-python==8.1.0
Werkzeug==2.3.7
email-validator==2.1.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
uuid==1.30
Pillow==10.1.0
python-slugify==8.0.1
